zik2ctl: show default paired Zik2 configuration with no arguments
This commit changes the default behavior of zik2ctl. Now, when no
arguments are given, it tries to connect to the default Zik2 device and
show its configuration. Using --device option each time to display and
configure Zik2 is painful as there will be likely only one Zik2 paired.

As a consequence, a new option appears to list paired devices: '-l'. In
case multiple Zik2 are paired with host, the device shall be selected
using the --device option.

Also uppercase all options descriptions.